Moved trim and extend into Network.tools.topology...this is broken
