+++ OpenPNM/Network/__GenericNetwork__.py
@@ -606,196 +606,6 @@ class GenericNetwork(Core):
         # Any existing adjacency and incidence matrices will be invalid
         self._update_network()
 
-    def extend(self,pore_coords=[],throat_conns=[],labels=[]):
-        r'''
-        Add individual pores (or throats) to the network from a list of coords
-        or conns.
-
-        Parameters
-        ----------
-        pore_coords : array_like
-            The coordinates of the pores to add
-        throat_conns : array_like
-            The throat connections to add
-        labels : string, or list of strings, optional
-            A list of labels to apply to the new pores and throats
-
-        Notes
-        -----
-        This needs to be enhanced so that it increases the size of all pore
-        and throat props and labels on ALL associated objects.  At the moment
-        if throws an error is there are ANY associated objects.
-
-        '''
-        if (self._geometries != []):
-            raise Exception('Network has active Geometries, cannot proceed')
-        if (self._phases != []):
-            raise Exception('Network has active Phases, cannot proceed')
-
-        logger.debug('Extending network')
-        Np_old = self.num_pores()
-        Nt_old = self.num_throats()
-        Np = Np_old + int(sp.size(pore_coords)/3)
-        Nt = Nt_old + int(sp.size(throat_conns)/2)
-        #Adjust 'all' labels
-        del self['pore.all'], self['throat.all']
-        self['pore.all'] = sp.ones((Np,),dtype=bool)
-        self['throat.all'] = sp.ones((Nt,),dtype=bool)
-        #Add coords and conns
-        if pore_coords != []:
-            coords = sp.vstack((self['pore.coords'],pore_coords))
-            self['pore.coords'] = coords
-        if throat_conns != []:
-            conns = sp.vstack((self['throat.conns'],throat_conns))
-            self['throat.conns'] = conns
-        for item in self.keys():
-            if item.split('.')[1] not in ['coords','conns','all']:
-                if item.split('.')[0] == 'pore':
-                    N = Np
-                else:
-                    N = Nt
-                if self[item].dtype == bool:
-                    temp = self[item]
-                    self[item] = sp.zeros((N,),dtype=bool)
-                    self[item][temp] = True
-                elif self[item].dtype == object:
-                    temp = self[item]
-                    self[item] = sp.ndarray((N,),dtype=object)
-                    self[item][sp.arange(0,sp.shape(temp)[0])] = temp
-                else:
-                    temp = self[item]
-                    try:
-                        self[item] = sp.ones((N,sp.shape(temp)[1]),dtype=float)*sp.nan
-                    except:
-                        self[item] = sp.ones((N,),dtype=float)*sp.nan
-                    self[item][sp.arange(0,sp.shape(temp)[0])] = temp
-        #Apply labels, if supplied
-        if labels != []:
-            #Convert labels to list if necessary
-            if type(labels) is str:
-                labels = [labels]
-            for label in labels:
-                #Remove pore or throat from label, if present
-                label = label.split('.')[-1]
-                if pore_coords != []:
-                    Ps = sp.r_[Np_old:Np]
-                    if 'pore.'+label not in self.labels():
-                        self['pore.'+label] = False
-                    self['pore.'+label][Ps] = True
-                if throat_conns != []:
-                    Ts = sp.r_[Nt_old:Nt]
-                    if 'throat.'+label not in self.labels():
-                        self['throat.'+label] = False
-                    self['throat.'+label][Ts] = True
-
-        self._update_network()
-
-    def trim(self, pores=[], throats=[]):
-        '''
-        Remove pores (or throats) from the network.
-
-        Parameters
-        ----------
-        pores (or throats) : array_like
-            A boolean mask of length Np (or Nt) or a list of indices of the
-            pores (or throats) to be removed.
-
-        Notes
-        -----
-        Trimming only adjusts Phase, Geometry, and Physics objects. Trimming a
-        Network that has already been used to run simulations will break those
-        simulation objects.
-
-        Examples
-        --------
-        >>> import OpenPNM
-        >>> pn = OpenPNM.Network.TestNet()
-        >>> pn.Np
-        125
-        >>> pn.Nt
-        300
-        >>> pn.trim(pores=[1])
-        >>> pn.Np
-        124
-        >>> pn.Nt
-        296
-
-        '''
-        for net in self.controller.networks():
-            if net._parent is self:
-                raise Exception('This Network has been cloned, cannot trim')
-
-        if pores != []:
-            pores = sp.array(pores,ndmin=1)
-            Pkeep = sp.ones((self.num_pores(),),dtype=bool)
-            Pkeep[pores] = False
-            Tkeep = sp.ones((self.num_throats(),),dtype=bool)
-            Ts = self.find_neighbor_throats(pores)
-            if len(Ts)>0:
-                Tkeep[Ts] = False
-        elif throats != []:
-            throats = sp.array(throats,ndmin=1)
-            Tkeep = sp.ones((self.num_throats(),),dtype=bool)
-            Tkeep[throats] = False
-            Pkeep = self['pore.all'].copy()
-        else:
-            logger.warning('No pores or throats recieved')
-            return
-
-        # Trim all associated objects
-        for item in self._geometries+self._physics+self._phases:
-            Pnet = self['pore.'+item.name]*Pkeep
-            Tnet = self['throat.'+item.name]*Tkeep
-            temp = self.map_pores(pores=sp.where(Pnet)[0],target=item,return_mapping=True)
-            Ps = temp['target']
-            temp = self.map_throats(throats=sp.where(Tnet)[0],target=item,return_mapping=True)
-            Ts = temp['target']
-            # Then resize 'all
-            item.update({'pore.all' : sp.ones((sp.sum(Pnet),),dtype=bool)})
-            item.update({'throat.all' : sp.ones((sp.sum(Tnet),),dtype=bool)})
-            # Overwrite remaining data and info
-            for key in list(item.keys()):
-                if key.split('.')[1] not in ['all']:
-                    temp = item.pop(key)
-                    if key.split('.')[0] == 'throat':
-                        logger.debug('Trimming {a} from {b}'.format(a=key,b=item.name))
-                        item[key] = temp[Ts]
-                    if key.split('.')[0] == 'pore':
-                        logger.debug('Trimming {a} from {b}'.format(a=key,b=item.name))
-                        item[key] = temp[Ps]
-
-        #Remap throat connections
-        Pmap = sp.ones((self.Np,),dtype=int)*-1
-        Pmap[Pkeep] = sp.arange(0,sp.sum(Pkeep))
-        tpore1 = self['throat.conns'][:,0]
-        tpore2 = self['throat.conns'][:,1]
-        Tnew1 = Pmap[tpore1[Tkeep]]
-        Tnew2 = Pmap[tpore2[Tkeep]]
-        #Write 'all' label specifically
-        self.update({'throat.all' : sp.ones((sp.sum(Tkeep),),dtype=bool)})
-        self.update({'pore.all' : sp.ones((sp.sum(Pkeep),),dtype=bool)})
-        # Write throat connections specifically
-        self.update({'throat.conns' : sp.vstack((Tnew1,Tnew2)).T})
-        # Overwrite remaining data and info
-        for item in list(self.keys()):
-            if item.split('.')[-1] not in ['conns','all']:
-                temp = self.pop(item)
-                if item.split('.')[0] == 'throat':
-                    logger.debug('Trimming {a} from {b}'.format(a=item,b=self.name))
-                    self[item] = temp[Tkeep]
-                if item.split('.')[0] == 'pore':
-                    logger.debug('Trimming {a} from {b}'.format(a=item,b=self.name))
-                    self[item] = temp[Pkeep]
-
-        #Reset network graphs
-        self._update_network(mode='regenerate')
-
-        #Check Network health
-        health = self.check_network_health()
-        if health['trim_pores'] != []:
-            logger.warning('Isolated pores exist!  Run check_network_health to ID which pores to remove.')
-            pass
-
     def stitch(self,donor,pores_1,pores_2,method='delaunay',len_max=sp.inf,label_suffix=''):
         r'''
         Stitches a second a network to the current network.
