+++ OpenPNM/Network/tools/__init__.py
@@ -8,4 +8,4 @@ Contents
 
 """
 
-import *
+from . import topology
